Fixed do_install_append so command doesn't run for shadow-native.

+++ recipes-extended/shadow/shadow_4.1.4.3.bbappend
@@ -5,5 +5,10 @@ FILESEXTRAPATHS_prepend := "${THISDIR}/${PN}-${PV}:"
 SRC_URI += "file://login.selinux"
 
 do_install_append() {
-	install -m 0644 ${WORKDIR}/login.selinux ${D}${sysconfdir}/pam.d/login
+	# If shadown-native is building the destination will not exist.  Now that shadow-native is
+	# a BBCLASSEXTEND this file applies to it also.  Not sure if there is a better way to 
+	# handle this.
+	if [ -e ${D}${sysconfdir}/pam.d ]; then
+		install -m 0644 ${WORKDIR}/login.selinux ${D}${sysconfdir}/pam.d/login
+	fi
 }
